+++ changelog.md
@@ -2,6 +2,7 @@
 ---
 * LLT-1113: Add no-link detection mechanism
 * LLT-4154: Optimize Upnp/Stun providers
+* LLT-4855: Add Aggregator struct for connectivity events gathering
 
 <br>